+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+No changes here yet.
+
+## [1.2.0] - 2024-12-17
+
 ### Deprecated
 - `TrackyMouse.cleanupDwellClicking()` is deprecated in favor of calling `dispose()` on the object returned by `TrackyMouse.initDwellClicking()`.
 
@@ -64,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Electron](https://electronjs.org/) app for desktop (not yet packaged for distribution).
 
 
-[Unreleased]: https://github.com/1j01/tracky-mouse/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/1j01/tracky-mouse/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/1j01/tracky-mouse/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/1j01/tracky-mouse/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/1j01/tracky-mouse/releases/tag/v1.0.0